Change parameters to hiera5

+++ modules/seges/manifests/packages.pp
@@ -1,27 +1,31 @@
-# Seges packages puppet class
+# Class: seges::packages
+# ===========================
+#
+# Class to manage packages installation
+#
 class seges::packages (
-  $packages_list   = $seges::params::packages_list,
-  $install_options = $seges::params::packages_install_options,
-) inherits seges::params {
-  exec { 'remove-old-fusioninventory':
-    path    => [ '/bin', '/sbin', '/usr/bin', '/usr/sbin' ],
-    command => 'yum -y remove fusioninventory-agent*',
-    onlyif  => [ 'test "`rpm -qi fusioninventory-agent | grep -o Vendor.*guillomovitch`"' ],
-  }
-
-  if $::osfamily == 'RedHat' {
-    class { 'seges_epel':
-      install_options => $install_options,
+  $packages_list,
+  $install_options,
+) {
+  if $::operatingsystem == 'CentOS' {
+    class { 'epel':
+      epel_mirrorlist => "https://mirrors.fedoraproject.org/mirrorlist?repo=epel-${::operatingsystemmajrelease}&arch=\$basearch&country=US",
     }
-  }
 
-  #seges::rpmforgerepo { 'rpmforge-release':
-  #  enable          => false,
-  #  install_options => $packages_install_options,
-  #} ->
+    exec { 'remove-old-fusioninventory':
+      path    => [ '/bin', '/sbin', '/usr/bin', '/usr/sbin' ],
+      command => 'yum -y remove fusioninventory-agent*',
+      onlyif  => [ 'test "`rpm -qa | grep fusioninventory-agent`"', 'test "`rpm -qi fusioninventory-agent | grep -o Vendor.*guillomovitch`"' ],
+    }
+    
+    package { ['nss-pam-ldapd', 'pam_ldap']: 
+      ensure => absent,
+    }
+
+  }
 
   package { $packages_list:
-    ensure  => installed,
+    ensure          => installed,
+    install_options => $install_options,
   }
-
 }
